src/requirements.txt
python-telegram-bot==20.0a4
python-dotenv
# logzero
structlog